refactor model files
