+++ server/routes/coinsRoutes.js
@@ -2,12 +2,6 @@ const coinsController = require("../controllers/coinsController");
 const express = require("express");
 const router = express.Router();
 const {checkAuth} = require("../services/middleware");
-/**
- * Route - coins/favourite
- * @summary - POST signin request of user
- */
-router.get("/favourite",checkAuth, coinsController.getFavourite);
-router.post("/favourite",checkAuth, coinsController.postFavourite);
 
 /**
  * Route - coins/buy
